Add methods to test for correct device, and add check for disk size

+++ src/main/perl/File.pm
@@ -55,11 +55,15 @@ sub _initialize
 Creates the file, with the desired name, size ownership and
 permissions.
 
+Returns 0 on success.
+
 =cut
 
 sub create
 {
 	my $self = shift;
+
+    return 1 if (! $self->is_correct_device);
 
 	# Don't overwrite existing files. Data loss and symlink
 	# attacks!
@@ -81,13 +85,40 @@ sub create
 
 Removes the file.
 
+Returns 0 on success.
+
 =cut
 
 sub remove
 {
 	my $self = shift;
 
+    return 1 if (! $self->is_correct_device);
+
 	return unlink ($self->devpath);
+}
+
+=pod
+
+=head2 devexists
+
+Returns true if the file exists in the system.
+
+=cut
+
+
+sub devexists
+{
+    my $self = shift;
+    return -f $self->devpath;
+}
+
+# Returns size in byte (assumes devpath exists).
+# Is used by size
+sub _size_in_byte
+{
+    my $self = shift;
+    return -s $self->devpath;
 }
 
 =pod
